backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
vaderSentiment==3.3.2
python-dotenv==1.0.0
tweepy==4.14.0
requests==2.31.0